Update assets URL and index sorting
It is 2022 and all the sudden I realized that "pandoc-slides" was
misleading to say the least, so I have update the GH name and because of
that I need to update the presentations' URL accordingly.

BTW, I am planning to make a little bit more interesting the "main"
index file, but for the moment I forced it to display the slide list in
a semi-chronological order.

Signed-off-by: Jonnatan Jossemar Cordero <jossemargt@gmail.com>