Update Talk index page styles
I wasn't satisfied by how the index looked like, so I invested a couple
of cycles improving its style. And to be frank I was trying to do
something more "elegant", however it was taking too much time, so I
preferred to stick with something in the middle.

For the record, I wanted to achieve something like @alvarotrigo shared
on this codepen snippet:

  https://codepen.io/alvarotrigo/pen/MWEgejK

Signed-off-by: Jonnatan Jossemar Cordero <jossemargt@gmail.com>