Change directory layout for gh-pages
I realized that typing that extra "static/" at the end of the URL was a
little bit annoying and misleading, even more when GH pages already have
that covered if and only if you use the "docs/" directory instead. So I
bite the bullet and updated the directory layout to comply with it.

Yeap, the search engines will penalize this abrupt change but meh. What
is SEO anyways? Am I right?

Signed-off-by: Jonnatan Jossemar Cordero <jossemargt@gmail.com>